+++ GPU/GPUTracking/TPCClusterFinder/GPUTPCClusterFinderKernels.h
@@ -18,7 +18,7 @@
 #include "GPUConstantMem.h"
 #include "GPUTPCSharedMemoryData.h"
 
-#include "cl/clusterFinderDefs.h"
+#include "clusterFinderDefs.h"
 
 namespace GPUCA_NAMESPACE
 {
